integrar vue y axios en vista tweets, eliminado sub menu-navbar( crear tweet), add function index en EntryController, add autorize en update y delete en EntryController, eliminados componentes inutilizados, y migrations basura, add router-web los resource api, agregado content en varible fillable User model

+++ app/Http/Controllers/EntryController.php
@@ -13,6 +13,11 @@ class EntryController extends Controller
     	$this->middleware('auth');
     }
 
+    public function index(){
+        return  Tweet::with('user')
+        ->where('user_id', auth()->id())->get();
+    }
+
     public function create(){
     	return view("entries.create");
     }
@@ -59,9 +64,9 @@ class EntryController extends Controller
 
     public function delete($id)
     {
-                //
+        $this->authorize('delete', $id);
         Tweet::find($id)->delete();
-        return back()->with(compact('status'));
+        
     }
 
     }
